+++ execution-core/src/transfer/transaction/phoenix.rs
@@ -97,6 +97,7 @@ impl Transaction {
                         hash: deploy.bytecode.hash,
                         bytes: Vec::new(),
                     },
+                    nonce: deploy.nonce,
                 })),
             },
             self.proof(),